+++ firehydrant/types.go
@@ -258,3 +258,40 @@ type Pagination struct {
 	Prev  int `json:"prev,omitempty"`
 	Next  int `json:"next,omitempty"`
 }
+
+// FunctionalityResponse is the payload for a single environment
+// URL: GET https://api.firehydrant.io/v1/functionalities/{id}
+type FunctionalityResponse struct {
+	ID          string                 `json:"id"`
+	Name        string                 `json:"name"`
+	Description string                 `json:"description"`
+	Slug        string                 `json:"slug"`
+	Services    []FunctionalityService `json:"services"`
+	Labels      map[string]string      `json:"labels"`
+	CreatedAt   time.Time              `json:"created_at"`
+	UpdatedAt   time.Time              `json:"updated_at"`
+}
+
+// FunctionalityService represents a service when creating a functionality
+type FunctionalityService struct {
+	ID string `json:"id"`
+}
+
+// CreateFunctionalityRequest is the payload for creating a service
+// URL: POST https://api.firehydrant.io/v1/services
+type CreateFunctionalityRequest struct {
+	Name        string                 `json:"name"`
+	Description string                 `json:"description"`
+	Services    []FunctionalityService `json:"services,omitempty"`
+	Labels      map[string]string      `json:"labels"`
+}
+
+// UpdateFunctionalityRequest is the payload for updating a environment
+// URL: PATCH https://api.firehydrant.io/v1/environments/{id}
+type UpdateFunctionalityRequest struct {
+	Name                    string                 `json:"name,omitempty"`
+	Description             string                 `json:"description"`
+	RemoveRemainingServices bool                   `json:"remove_remaining_services"`
+	Labels                  map[string]string      `json:"labels"`
+	Services                []FunctionalityService `json:"services"`
+}